explicitly convert `MathLib::big{u}int` to other types
fixes some compiler warnings with `USE_BOOST_INT128` as those implicit conversions are not allowed

+++ lib/mathlib.cpp
@@ -116,7 +116,7 @@ void MathLib::value::promote(const MathLib::value &v)
         }
     } else if (!isFloat()) {
         mIsUnsigned = false;
-        mDoubleValue = mIntValue;
+        mDoubleValue = (double)mIntValue;
         mType = MathLib::value::Type::FLOAT;
     }
 }
